+++ coverage-tests/configuration/index.js
@@ -1,4 +1,3 @@
-
 module.exports = {
   name: 'eyes_selenium_java',
   emitter: 'https://raw.githubusercontent.com/applitools/sdk.coverage.tests/master/java/emitter.js',