+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## [vNext]
+## [3.192.0]
 ### Update
 - Cleanup cached data after check() execution. [Trello 1673](https://trello.com/c/CYbkzXia)
 - Using original view location on the screen for the Android scroll. [Trello 2138](https://trello.com/c/EI2mxmTD)